src/requirements.txt
certifi==2025.8.3
babel==2.17.0
flask-babel==4.0.0
flask==3.1.1
jinja2==3.1.6
lxml==6.0.0
pygments==2.19.2
python-dateutil==2.9.0.post0
pyyaml==6.0.2
httpx[http2]==0.28.1
httpx-socks[asyncio]==0.10.0
Brotli==1.1.0
uvloop==0.21.0
setproctitle==1.3.6
valkey==6.1.0
markdown-it-py==3.0.0
fasttext-predict==0.9.2.4
tomli==2.2.1; python_version < '3.11'
msgspec==0.19.0
typer-slim==0.16.0
isodate==0.7.2
whitenoise==6.9.0
azure-functions==1.18.0
packaging==24.1
requests==2.32.4
